CBD-2044: Initial refactoring for shared Python packages
In order to allow the shared Python packages originally in build-infra
to be installed on their own, they have been moved to their own repository
with a setup/requirements configuration added.  The following packages
were moved:

  - manifest
  - util (renamed to cbutil)

During the migration a new version of Dulwich came out which has a fix
for the 'origin' bug for porcelain.clone, so the cbutil/git.py file was
modified to simplify the RepoCache code for the fix.

Change-Id: If89b5537fdc6fad26f2a50791d8379e21ab295aa
Reviewed-on: http://review.couchbase.org/85084
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>