CBD-2183: Add checkout_repo() method to cbutil.git module
A new method, checkout_repo(), which will take a given path and URL
and either clone the given URL to the path if it's not there or fetch
the latest data if it is.

The fetch() method originally in the RepoCache class has been moved
to the top-level of the module for shared use.

Change-Id: I7de977b48ebfda04037fa6320fcd09ca1972c6e6
Reviewed-on: http://review.couchbase.org/85816
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>