CBD-2282: Some code generalization for future use
A few of the query methods were combined into a single, generalized method
to allow for future expansion and using a generator to allow iteration over
the results (instead of a large list being returned) for efficiency reasons.

Also corrected the duplicate 'get_commit()' methods and generalized it as
well to allow for either a document key or a project and SHA pair.

Change-Id: Ia756304203835a582ba56e0c303cd8e3b17f247c
Reviewed-on: http://review.couchbase.org/88289
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>