CBD-2301: Correct flaw in get_repo() for cache
The get_repo() method in RepoCache in cbutil.git had a serious flaw where
the remotes weren't getting properly added and certain checks were not
being done to insure the right URL was being pulled from.  This new code
corrects that and now assumes each remote will have a single URL (the
most recent seen in the manifests) instead of a list, which will require
changes anywhere this was used.

Change-Id: I1fd6f38bc02cd0a0acb1cfc2f3baa63871e479d9
Reviewed-on: http://review.couchbase.org/88677
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>