CBD-2729: Fix remote add in cbbuild.cbutil's git module
For cbbuild.cbutil.git, new remote entries for a given existing repository
were added without a 'fetch' line (refspec) in the Git config file (bare
repository).  This led to failures when attempting to check out a certain
branch that resulted in bad data in the build database.  This change will
fix the issue.

Note that dulwich has been locked to the currently used version to avoid
any issues with more recent releases.

Change-Id: Ibb4a4054890e00460130716170319b14bfe40901
Reviewed-on: http://review.couchbase.org/100540
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>