CBD-2759: Use LOCKMODE_WAIT for bucket handling in cbutil.db
To resolve an issue with multiple threads/connections coming from
the REST API, the 'lockmode' for the bucket creation in the cbutil.db
module was set to LOCKMODE_WAIT, which should prevent the error that
was being seen from occurring again.

Note that this may have a bit of a performance hit; if it does prove
to be a stumbling block down the road, will need to look at either
txcouchbase or acouchbase for asynchronous handling of the connections,
but this should unstick things for now.

Change-Id: I56a116a7387f583e74e877edbfc70a4c48524d9e
Reviewed-on: http://review.couchbase.org/101458
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>